Moved IDrawingContextImpl to Avalonia.Platform.
So it can be alongside its impl brethren.
