+++ src/Avalonia.Visuals/Platform/IDrawingContextImpl.cs
@@ -2,9 +2,10 @@
 // Licensed under the MIT license. See licence.md file in the project root for full license information.
 
 using System;
+using Avalonia.Media;
 using Avalonia.Platform;
 
-namespace Avalonia.Media
+namespace Avalonia.Platform
 {
     /// <summary>
     /// Defines the interface through which drawing occurs.
